fix(랭킹페이지): myRank int -> Long #85
h2에서는 됐지만 mysql에서는 안되는 현상

+++ src/main/java/com/hh99/nearby/rank/service/RankPageService.java
@@ -49,7 +49,7 @@ public class RankPageService {
             for (int i = 0; i < rankPageDtos.size(); i++) {
                 if (user.getUsername().equals(rankPageDtos.get(i).getNickname())) {
                     System.out.println(i + 1);
-                    int myRank = i+1;
+                    long myRank = (long)i+1;
                     String nickname = rankPageDtos.get(i).getNickname();
                     memberRepository.updateRank(myRank, nickname);
                 }
